+++ lib/Memcached.php
@@ -123,8 +123,7 @@ class Memcached extends BaseCache
         } else {
             $params = array($keysWithPrefix, $cas, \Memcached::GET_PRESERVE_ORDER);
         }
-        $values = call_user_func_array(array($this->object, 'getMulti'), $params);
-
+        $values = (array) call_user_func_array(array($this->object, 'getMulti'), $params);
         return array_combine($keys, $values);
     }